extract out config class
